# Conflicts:
#   client/src/main/java/client/controllers/MainScreenActivityCtrl.java
#   client/src/main/java/client/controllers/SceneCtrl.java

+++ client/src/main/java/client/controllers/QuestionActivityCtrl.java
@@ -54,7 +54,7 @@ public class QuestionActivityCtrl {
         this.sampleQuestion.setText("How much does electricity(in kWH) does " +
                 question.getTitle().substring(0, 1).toLowerCase(Locale.ROOT) + question.getTitle().substring(1) + " take?");
 
-        int answer = question.getConsumption();
+        long answer = (question.getConsumption());
         long range = Math.round(Math.random() * answer);
         long range2 = Math.round(Math.random() * answer);
         double random = Math.random() * 3;
